[BPK-2804] Implemented the feature.
