+++ Backpack/src/main/java/net/skyscanner/backpack/dialog/BpkDialog.kt
@@ -6,7 +6,6 @@ import android.graphics.Color
 import net.skyscanner.backpack.R
 import net.skyscanner.backpack.text.BpkText
 import android.graphics.drawable.ColorDrawable
-import android.graphics.drawable.GradientDrawable
 import android.util.DisplayMetrics
 import android.view.ViewGroup
 import android.view.Window
@@ -15,8 +14,9 @@ import android.view.LayoutInflater
 import android.view.View
 import androidx.annotation.ColorRes
 import androidx.annotation.DrawableRes
-import androidx.appcompat.widget.LinearLayoutCompat
-import androidx.core.content.res.ResourcesCompat
+import net.skyscanner.backpack.dialog.internal.BpkDialogIcon
+import net.skyscanner.backpack.dialog.internal.DialogConstants
+import net.skyscanner.backpack.dialog.internal.DialogWindowLayout
 
 open class BpkDialog(
   context: Context,
@@ -55,13 +55,7 @@ open class BpkDialog(
     }
 
   fun addActionButton(button: View) {
-    button.layoutParams = LinearLayoutCompat.LayoutParams(
-      LinearLayoutCompat.LayoutParams.MATCH_PARENT,
-      LinearLayoutCompat.LayoutParams.WRAP_CONTENT
-    ).apply {
-      setMargins(0, 0, 0, variables.buttonStackItemSpace)
-    }
-    viewHolder?.contentLayout?.addView(button)
+    viewHolder?.buttonsRoot?.addView(button, ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT)
   }
 
   private fun setupDialog() {
@@ -75,10 +69,10 @@ open class BpkDialog(
     window?.setLayout(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT)
     window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
     if (style == Style.BOTTOM_SHEET) {
-      viewHolder?.container?.verticalGravity = DialogContentLayout.Gravity.Bottom
+      viewHolder?.container?.verticalGravity = DialogWindowLayout.Gravity.Bottom
       window?.setWindowAnimations(R.style.Bpk_dialog_animation)
     } else {
-      viewHolder?.container?.verticalGravity = DialogContentLayout.Gravity.Center
+      viewHolder?.container?.verticalGravity = DialogWindowLayout.Gravity.Center
     }
     viewHolder?.container?.dismissListener = {
       dismiss()
@@ -93,34 +87,15 @@ open class BpkDialog(
     val inflater = context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
     val view = inflater.inflate(R.layout.bpk_dialog, null)
 
-    viewHolder = ViewHolder((view as ViewGroup).getChildAt(0) as DialogContentLayout).apply {
-      val params = contentLayout.layoutParams as ViewGroup.MarginLayoutParams
-      params.topMargin = variables.contentContainerMarginTop
-      contentLayout.background = getContentBackground()
-      contentLayout.setPadding(
-        variables.contentContainerPaddingHorizontal,
-        variables.contentContainerPaddingTop,
-        variables.contentContainerPaddingHorizontal,
-        variables.contentContainerPaddingBottom
-      )
-    }
+    viewHolder = ViewHolder((view as ViewGroup).getChildAt(0) as DialogWindowLayout)
 
     return view
   }
 
-  private fun getContentBackground(): GradientDrawable {
-    val shape = GradientDrawable()
-    shape.shape = GradientDrawable.RECTANGLE
-    val radii = variables.contentRadius
-    shape.cornerRadii = floatArrayOf(radii, radii, radii, radii, radii, radii, radii, radii)
-    shape.setColor(ResourcesCompat.getColor(context.resources, R.color.bpkWhite, context.theme))
-    return shape
-  }
-
-  private class ViewHolder(val container: DialogContentLayout) {
+  private class ViewHolder(val container: DialogWindowLayout) {
     val title: BpkText = container.findViewById(R.id.dialog_title)
     val description: BpkText = container.findViewById(R.id.dialog_description)
     val iconView: BpkDialogIcon = container.findViewById(R.id.dialog_icon)
-    val contentLayout: ViewGroup = container.findViewById(R.id.dialog_content_layout)
+    val buttonsRoot: ViewGroup = container.findViewById(R.id.buttons_root_layout)
   }
 }
