+++ Backpack/src/main/java/net/skyscanner/backpack/dialog/internal/DialogInsetsLayout.kt
@@ -1,0 +1,16 @@
+package net.skyscanner.backpack.dialog.internal
+
+import android.content.Context
+import android.util.AttributeSet
+import android.widget.FrameLayout
+
+internal class DialogInsetsLayout @JvmOverloads constructor(
+  context: Context,
+  attrs: AttributeSet? = null,
+  defStyleAttr: Int = 0
+) : FrameLayout(context, attrs, defStyleAttr) {
+
+  init {
+    fitsSystemWindows = true
+  }
+}
