+++ Backpack/src/main/res/drawable/bpk_bg_dialog_bottom.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+  <corners android:bottomLeftRadius="@dimen/bpkBorderRadiusSm" android:bottomRightRadius="@dimen/bpkBorderRadiusSm" />
+  <solid android:color="@color/bpk_dialog_background" />
+  <padding android:left="@dimen/bpk_dialog_padding_horizontal" android:right="@dimen/bpk_dialog_padding_horizontal" />
+</shape>
