+++ Backpack/src/main/res/drawable/bpk_bg_dialog_divider.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+  <size android:height="@dimen/bpk_dialog_spacing_button" />
+</shape>
