+++ Backpack/src/main/res/layout/bpk_dialog.xml
@@ -1,66 +1,71 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<net.skyscanner.backpack.dialog.internal.DialogInsetsLayout xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:app="http://schemas.android.com/apk/res-auto"
   android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  android:fitsSystemWindows="true">
+  android:layout_height="match_parent">
 
-  <net.skyscanner.backpack.dialog.DialogContentLayout
+  <net.skyscanner.backpack.dialog.internal.DialogWindowLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <ScrollView
-      android:layout_width="0dp"
+    <androidx.appcompat.widget.LinearLayoutCompat
+      android:layout_width="match_parent"
       android:layout_height="wrap_content"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      app:layout_constraintWidth_max="580dp">
+      android:orientation="vertical"
+      app:layout_constraintWidth_max="@dimen/abc_config_prefDialogWidth">
 
-      <RelativeLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content">
+      <ScrollView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:scrollbarStyle="outsideOverlay">
 
         <androidx.appcompat.widget.LinearLayoutCompat
-          android:id="@+id/dialog_content_layout"
           android:layout_width="match_parent"
           android:layout_height="wrap_content"
-          android:layout_alignParentTop="true"
-          android:layout_centerHorizontal="true"
-          android:background="@color/bpkWhite"
-          android:orientation="vertical" >
+          android:background="@drawable/bpk_bg_dialog_top"
+          android:orientation="vertical">
+
+          <net.skyscanner.backpack.dialog.internal.BpkDialogIcon
+            android:id="@+id/dialog_icon"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center_horizontal" />
 
           <net.skyscanner.backpack.text.BpkText
             android:id="@+id/dialog_title"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:layout_marginTop="@dimen/bpk_dialog_spacing_title"
             android:textAlignment="center"
             app:textStyle="xl"
             app:weight="emphasized" />
 
           <net.skyscanner.backpack.text.BpkText
             android:id="@+id/dialog_description"
-            android:layout_marginTop="@dimen/bpkSpacingMd"
-            android:layout_marginBottom="@dimen/bpkSpacingLg"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:layout_marginTop="@dimen/bpkSpacingMd"
+            android:layout_marginBottom="@dimen/bpkSpacingLg"
             android:textAlignment="center"
             app:textStyle="lg" />
 
         </androidx.appcompat.widget.LinearLayoutCompat>
 
-        <net.skyscanner.backpack.dialog.BpkDialogIcon
-          android:id="@+id/dialog_icon"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"
-          android:layout_alignParentTop="true"
-          android:layout_centerHorizontal="true" />
+      </ScrollView>
 
-      </RelativeLayout>
+      <androidx.appcompat.widget.LinearLayoutCompat
+        android:id="@+id/buttons_root_layout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@drawable/bpk_bg_dialog_bottom"
+        android:orientation="vertical"
+        android:paddingBottom="@dimen/bpk_dialog_padding_bottom"
+        app:divider="@drawable/bpk_bg_dialog_divider"
+        app:showDividers="middle|end" />
 
-    </ScrollView>
+    </androidx.appcompat.widget.LinearLayoutCompat>
 
-  </net.skyscanner.backpack.dialog.DialogContentLayout>
+  </net.skyscanner.backpack.dialog.internal.DialogWindowLayout>
 
-</FrameLayout>
+</net.skyscanner.backpack.dialog.internal.DialogInsetsLayout>
