+++ Backpack/src/main/res/values/colors.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+  <color name="bpk_dialog_background">@color/bpkWhite</color>
+</resources>
